+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - New #125: Add `PreparedInlineMessage` type.
 - New #125: Add `savePreparedInlineMessage` method.
 - New #125: Add `Gift` and `Gifts` types.
+- New #125: Add `getAvailableGifts` method.
 
 ## 0.4.5 October 31, 2024